app/requirements.txt
streamlit
Pillow==8.3.2
numpy==1.18.5
boto3==1.18.20